pyproject.toml
[project]
name = "pythonproject3"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=22.3",
]
